requirements.txt
streamlit
google-generativeai
Pillow
ics
python-dateutil
PyMuPDF
anthropic
icalendar